subscription/serializers.py
from rest_framework import serializers

from .models import Subscription


class SubscriptionSerializer(serializers.ModelSerializer):
    class Meta:
        model = Subscription
        fields = ['start_date', 'end_date', 'customer', 'service']
